Fixed UI, changed project structure, and added formatting for json and xml output
