+++ popup.html
@@ -35,6 +35,6 @@
         <button class="button" id="btn-copy-uri-escape">Copy</button>
     </div>
 
-    <script src="popup.js"></script>
+    <script type="module" src="src/popup.js"></script>
 </body>
 </html>
