+++ src/handlers/copy.js
@@ -1,0 +1,5 @@
+export function copyToClipboard(elementId) {
+    const copyText = document.getElementById(elementId);
+    copyText.select();
+    document.execCommand("copy");
+}
