+++ src/handlers/openTab.js
@@ -1,0 +1,13 @@
+export function openTab(tabId, buttonId) {
+    const tabs = document.querySelectorAll('.tab');
+    tabs.forEach(tab => {
+        tab.classList.remove('active');
+    });
+    document.getElementById(tabId).classList.add('active');
+
+    const tabButtons = document.querySelectorAll('.tab-option');
+    tabButtons.forEach(button => {
+        button.classList.remove('active');
+    });
+    document.getElementById(buttonId).classList.add('active');
+}
