+++ styles.css
@@ -45,7 +45,7 @@ h1 {
 }
 
 textarea {
-    width: 100%;
+    width: 95%;
     height: 100px;
     margin-bottom: 10px;
     padding: 10px;
@@ -59,6 +59,7 @@ textarea {
     display: inline-block;
     padding: 10px 20px;
     margin-right: 10px;
+    margin-bottom: 10px;
     border: none;
     border-radius: 5px;
     background-color: #5a67d8;
